# docs/swagger.yaml
basePath: /api/v2
definitions:
...
  handler.CreateNodeRequest:
    properties:
      api_port:
        type: integer
      api_token:
        type: string
      bandwidth:
        type: integer
      host:
        type: string
      isp:
        type: string
      max_conn:
        type: integer
      name:
        type: string
      port:
        maximum: 65535
        minimum: 1
        type: integer
      region:
        type: string
      type:
        enum:
        - relay
        - exit
        type: string
      weight:
        type: integer
    required:
    - host
    - name
    - port
    - type
    type: object
...
  handler.ToggleRequest:
    properties:
      enabled:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@anixops.com
    name: API Support
    url: https://github.com/AnixOps/anix-control/issues
  description: |-
    AnixOps Control 统一控制面 API 文档
    支持用户管理、节点管理、订阅系统、支付网关、流量转发等功能
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/AnixOps/anix-control
  title: AnixOps Control API
  version: 3.0.0-alpha.1
paths:
...
  /admin/forward/nodes:
    get:
      consumes:
      - application/json
      description: 管理员获取中转节点列表，支持分页和类型筛选
      parameters:
      - description: 节点类型 (relay/exit)
        in: query
        name: type
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取中转节点列表
      tags:
      - 管理端-转发
    post:
      consumes:
      - application/json
      description: 管理员创建新的中转节点
      parameters:
      - description: 节点信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateNodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建中转节点
      tags:
      - 管理端-转发
  /admin/forward/nodes/{id}:
    delete:
      consumes:
      - application/json
      description: 管理员删除指定中转节点
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除中转节点
      tags:
      - 管理端-转发
    get:
      consumes:
      - application/json
      description: 管理员获取指定中转节点的详细信息
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取中转节点详情
      tags:
      - 管理端-转发
    put:
      consumes:
      - application/json
      description: 管理员更新指定中转节点的信息
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: integer
      - description: 节点更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateNodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新中转节点
      tags:
      - 管理端-转发
  /admin/forward/nodes/{id}/check:
    post:
      consumes:
      - application/json
      description: 管理员对指定中转节点进行健康检查
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 健康检查节点
      tags:
      - 管理端-转发
  /admin/forward/nodes/{id}/sync-stats:
    post:
      consumes:
      - application/json
      description: 从 gost 节点同步流量和连接统计数据到面板
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 同步节点统计数据
      tags:
      - 管理端-转发
  /admin/forward/nodes/{id}/toggle:
    post:
      consumes:
      - application/json
      description: 管理员启用或禁用指定中转节点
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ToggleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 切换节点状态
      tags:
      - 管理端-转发
  /admin/forward/rules:
    get:
      consumes:
      - application/json
      description: 管理员获取转发规则列表，支持分页和用户筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 用户ID
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取转发规则列表
      tags:
      - 管理端-转发
    post:
      consumes:
      - application/json
      description: 管理员创建新的转发规则
      parameters:
      - description: 规则信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建转发规则
      tags:
      - 管理端-转发
  /admin/forward/rules/{id}:
    delete:
      consumes:
      - application/json
      description: 管理员删除指定转发规则
      parameters:
      - description: 规则ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除转发规则
      tags:
      - 管理端-转发
    get:
      consumes:
      - application/json
      description: 管理员获取指定转发规则的详细信息
      parameters:
      - description: 规则ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取转发规则详情
      tags:
      - 管理端-转发
    put:
      consumes:
      - application/json
      description: 管理员更新指定转发规则的信息
      parameters:
      - description: 规则ID
        in: path
        name: id
        required: true
        type: integer
      - description: 规则更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新转发规则
      tags:
      - 管理端-转发
  /admin/forward/rules/{id}/toggle:
    post:
      consumes:
      - application/json
      description: 管理员启用或禁用指定转发规则
      parameters:
      - description: 规则ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ToggleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 切换规则状态
      tags:
      - 管理端-转发
  /admin/forward/stats:
    get:
      consumes:
      - application/json
      description: 管理员获取转发统计数据，包括节点数、在线数、流量等
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取转发统计
      tags:
      - 管理端-转发
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...